refactor: presence mock과 참조 경로 구조 정리
